pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "authutils"
version = "0.1.0"
description = "A modular authentication library that provides easy token-based authentication"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Evan Vera", email = "evan.vera@hotmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "python-jose[cryptography]",
    "httpx",
    "pydantic",
    "cachetools"
]

[tool.hatch.build.targets.wheel]
packages = ["src/authutils"]